python/pyproject.toml
[project]
name = "resnet-18"
version = "0.1.0"
description = "ResNet-18 example"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["transformers", "torch", "datasets", "Pillow"]
